Testing commit after migration of repo from personal account to organization
DoubleComparer is just a modification of the FloatComparer class that I
made before FloatComparer had any features for handling double
precisions floating point numbers.  It may not be strictly necessary
anymore, but it's used in my existing plug-in projects and replacing it
hasn't bubbled to the top of my TODO list yet.